+++ Extract/krkr/FateFD.cpp
@@ -17,7 +17,7 @@ bool CFateFD::OnCheckDecrypt(CArcFile* archive)
 ///
 /// @param archive Archive
 ///
-DWORD CFateFD::OnInitDecrypt(CArcFile* archive)
+u32 CFateFD::OnInitDecrypt(CArcFile* archive)
 {
 	const SFileInfo* file_info = archive->GetOpenFileInfo();
 
@@ -74,7 +74,7 @@ DWORD CFateFD::OnInitDecrypt(CArcFile* archive)
 /// @param offset      Location of data to be decoded
 /// @param decrypt_key Decryption key
 ///
-DWORD CFateFD::OnDecrypt(BYTE* target, DWORD target_size, DWORD offset, DWORD decrypt_key)
+size_t CFateFD::OnDecrypt(u8* target, size_t target_size, size_t offset, u32 decrypt_key)
 {
 	for (size_t i = 0; i < target_size; i++)
 	{
