+++ Extract/krkr/SisMiko.h
@@ -6,6 +6,6 @@ class CSisMiko final : public CKrkr
 {
 public:
 	bool  OnCheckDecrypt(CArcFile* archive) override;
-	DWORD OnInitDecrypt(CArcFile* archive) override;
-	DWORD OnDecrypt(BYTE* target, DWORD target_size, DWORD offset, DWORD decrypt_key) override;
+	u32 OnInitDecrypt(CArcFile* archive) override;
+	size_t OnDecrypt(u8* target, size_t target_size, size_t offset, u32 decrypt_key) override;
 };
